+++ Runtime/Authentication.cs
@@ -3,24 +3,78 @@ using System.Threading.Tasks;
 using Auth;
 using Data;
 using UnityEngine;
+using UnityEngine.Networking;
 
 namespace YourePlugin
 {
     public class Authentication
     {
         public event Action<YoureUser> SignInSucceeded;
-        public event Action<string> SignInFailed;
+        public event Action SignInFailed;
 
         private readonly string _deeplinkScheme;
         private readonly string _authority;
         private readonly string _clientId;
         private AuthClient _authClient;
+        private bool _signInInProgress;
 
         public Authentication(string clientId,  string authority, string deeplinkScheme) 
         {
             _clientId = clientId;
             _authority = authority;
             _deeplinkScheme = $"{deeplinkScheme}://keycloak_callback";
+        }
+        
+        
+        /// <summary>
+        /// Will return TRUE if sign-in was used.
+        /// </summary>
+        public bool WasSignedIn()
+        {
+            string savedId = PlayerPrefs.GetString("YREID_lastID");
+            if (savedId == null)
+                return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Will return YoureUser Object if previous session is still valid
+        /// </summary>
+        public async Task<YoureUser> GetActiveUser()
+        {
+            string savedId = PlayerPrefs.GetString("YREID_lastID");
+            if (savedId == null)
+                return null;
+
+            string savedAccessToken = PlayerPrefs.GetString("YREID_lastAccessToken");
+            if(savedAccessToken == null)
+                return null;
+
+            UnityWebRequest request = UnityWebRequest.Get($"{_authority}/protocol/openid-connect/userinfo");
+            request.SetRequestHeader("Authorization", "Bearer " + savedAccessToken);
+            await request.SendWebRequest();
+
+            if (request.result == UnityWebRequest.Result.Success)
+            {
+                if (request.downloadHandler.text.Contains(savedId))
+                {
+                    YoureUser user = new()
+                    {
+                        Id = savedId,
+                        Email = PlayerPrefs.GetString("YREID_lastEmail"),
+                        UserName = PlayerPrefs.GetString("YREID_lastUserName"),
+                        AccessToken = savedAccessToken,
+                    };
+                    return user;
+                }
+            }
+            else
+            {
+                Youre.LogDebug("Old session not valid: " + request.error);
+            }
+
+            return null;
         }
 
         public async Task<bool> SignOut()
@@ -30,37 +84,70 @@ namespace YourePlugin
                 Youre.LogDebug("No user signed in");
                 return false;
             }
-            GameObject oldSignInGO = GameObject.Find("SignInCanvas");
-            if (oldSignInGO != null)
+
+            bool isSignedOut;
+            
+            if (Application.isMobilePlatform)
             {
-                UnityEngine.Object.Destroy(oldSignInGO);
+                GameObject oldSignInGO = GameObject.Find("SignInCanvas");
+                if (oldSignInGO != null)
+                    UnityEngine.Object.Destroy(oldSignInGO);
+                GameObject signIn = new GameObject("SignInCanvas");
+                MobileSignInBehavior go = signIn.AddComponent<MobileSignInBehavior>();
+                go.SignInFailed = () => SignInFailed?.Invoke();
+                isSignedOut = await go.SignOut(_authClient);
+                UnityEngine.Object.Destroy(signIn);
             }
-            GameObject signIn = new GameObject("SignInCanvas");
-            SignInBehavior go = signIn.AddComponent<SignInBehavior>();
-            go.SignInFailed = () => SignInFailed?.Invoke("failed");
-            bool signedOut = await go.SignOut(_authClient);
-            UnityEngine.Object.Destroy(signIn);
+            else
+                isSignedOut = await _authClient.LogoutAsync();
+          
+            PlayerPrefs.DeleteKey("YREID_lastAccessToken");
+            PlayerPrefs.DeleteKey("YREID_lastID");
+            PlayerPrefs.DeleteKey("YREID_lastEmail");
+            PlayerPrefs.DeleteKey("YREID_lastUserName");
             _authClient = null;
-            return signedOut;
+            
+            return isSignedOut;
         }
 
         public async Task SignIn() 
         {
+            if(_signInInProgress)
+                return;
+
+            
             _authClient = new AuthClient(_clientId, _authority, _deeplinkScheme);
-            GameObject oldSignInGO = GameObject.Find("SignInCanvas");
-            if (oldSignInGO != null)
+            Youre.LogDebug("Signing in...");
+            
+            AuthClientResult result = null;
+
+            _signInInProgress = true;
+            if (Application.isMobilePlatform)
             {
-                UnityEngine.Object.Destroy(oldSignInGO);
+                GameObject oldSignInGO = GameObject.Find("SignInCanvas");
+                if (oldSignInGO != null)
+                {
+                    UnityEngine.Object.Destroy(oldSignInGO);
+                }
+                GameObject signIn = new GameObject("SignInCanvas");
+                MobileSignInBehavior go = signIn.AddComponent<MobileSignInBehavior>();
+                result = await go.SignIn(_authClient);
+                UnityEngine.Object.Destroy(signIn);
             }
-            GameObject signIn = new GameObject("SignInCanvas");
-            SignInBehavior go = signIn.AddComponent<SignInBehavior>();
+            else
+            {
+                try
+                {
+                    result = await _authClient.LoginAsync();
+                }
+                catch (Exception e)
+                {
+                    Youre.LogDebug("error:");
+                    Youre.LogDebug(e.ToString());
+                }
+            }
             
-            go.SignInFailed = () => SignInFailed?.Invoke("failed");
-            go.SignInCancelled = () => SignInFailed?.Invoke("canceled");
-            
-            AuthClientResult result = await go.SignIn(_authClient);
-            
-            UnityEngine.Object.Destroy(signIn);
+            _signInInProgress = false;
             
             if (result != null)
             {
@@ -71,7 +158,17 @@ namespace YourePlugin
                     UserName = result.UserName,
                     AccessToken = result.AccessToken,
                 };
+                
+                PlayerPrefs.SetString("YREID_lastAccessToken", user.AccessToken);
+                PlayerPrefs.SetString("YREID_lastID", user.Id);
+                PlayerPrefs.SetString("YREID_lastEmail", user.Email);
+                PlayerPrefs.SetString("YREID_lastUserName", user.UserName);
+                
                 SignInSucceeded?.Invoke(user);
+            } 
+            else
+            {
+                SignInFailed?.Invoke();
             }
         }
     }